analytics/backend/pyproject.toml
[tool.poetry]
authors = ["Jan Gosmann <jan@hyper-world.de>"]
description = ""
license = "proprietary"
name = "cdm-analytics"
version = "0.1.0"

[tool.poetry.dependencies]
fastapi = "^0.73.0"
psycopg = "^3.0.8"
psycopg-binary = "^3.0.8"
psycopg2-binary = "^2.9.3"
python = "^3.8"
yoyo-migrations = "^7.3.2"
psycopg-pool = "^3.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
mypy = "^0.931"
pylint = "^2.12.2"
pytest = "^7.0.1"
requests = "^2.27.1"
uvicorn = "^0.17.4"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
